app/requirements.txt
PyYAML==5.1.2
psycopg2==2.8.3
tornado==6.1
tornado-sqlalchemy==0.7